workload_requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
_pytorch_select=0.1=cpu_0
audioread=2.1.9=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.8=ha878542_0
certifi=2021.10.8=py37h89c1867_2
chainmap=1.0.3=pypi_0
charset-normalizer=2.0.12=pypi_0
cheroot=8.6.0=pypi_0
cherrypy=11.2.0=pypi_0
click=8.1.3=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
decorator=5.1.1=pypi_0
ffmpeg=4.3=hf484d3e_0
flask=2.1.2=pypi_0
freetype=2.11.0=h70c0345_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
idna=3.3=pypi_0
importlib-metadata=4.11.3=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
itsdangerous=2.1.2=pypi_0
jaraco-classes=3.2.1=pypi_0
jaraco-functools=3.5.0=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.1.0=pypi_0
jpeg=9d=h7f8727e_0
kaldi_io=0.9.4=pyhd8ed1ab_0
kaldiio=2.17.2=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
librosa=0.6.3=pypi_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
llvmlite=0.31.0=pypi_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.1.1=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
more-itertools=8.12.0=pypi_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=py37hd09550d_3
numba=0.48.0=pypi_0
numpy=1.21.5=py37he7a7128_1
numpy-base=1.21.5=py37hf524024_1
olefile=0.46=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1n=h7f8727e_0
pillow=4.3.0=pypi_0
pip=21.2.2=py37h06a4308_0
portaudio=19.6.0=h7b6447c_4
portend=3.1.0=pypi_0
pyaudio=0.2.11=py37h7b6447c_2
pyopengl=3.1.6=pypi_0
pyopengl-accelerate=3.1.5=pypi_0
python=3.7.13=h12debd9_0
python_abi=3.7=2_cp37m
pytorch=1.10.1=py3.7_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
pytorch-pcen=0.0.1=pypi_0
pyttsx3=2.6=pypi_0
pytz=2022.1=pypi_0
pyyaml=5.3.1=py37h7b6447c_1
readline=8.1.2=h7f8727e_1
requests=2.27.1=pypi_0
resampy=0.2.2=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
setuptools=61.2.0=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
speechrecognition=3.7.1=pypi_0
sqlite=3.38.2=hc218d9a_0
tempora=5.0.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.11=h1ccaba5_0
torchvision=0.11.2=py37_cu102
tqdm=4.64.0=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.9=pypi_0
werkzeug=2.1.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zipp=3.8.0=pypi_0
zlib=1.2.12=h7f8727e_2
zstd=1.4.9=haebb681_0
